gonullu-ai-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.0
aiosqlite
asyncpg==0.30.0
psycopg2-binary==2.9.9
pydantic[email]==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.10
redis==5.0.8
celery==5.4.0
slowapi==0.1.9
boto3==1.35.0
httpx==0.27.0
reportlab==4.2.0
python-dotenv==1.0.1
itsdangerous==2.2.0
